Moved Menus and DefinitionBlock
