+++ Code/src/main/java/nl/utwente/group10/ui/components/menu/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Contains components for the menus that can be used to visually program a Haskell application.
  */
-package nl.utwente.group10.ui.menu;
+package nl.utwente.group10.ui.components.menu;
