+++ Code/src/main/resources/ui/FunctionMenu.fxml
@@ -3,7 +3,7 @@
 <?import javafx.scene.layout.FlowPane?>
 <?import javafx.scene.layout.VBox?>
 
-<fx:root type="nl.utwente.group10.ui.menu.FunctionMenu" xmlns:fx="http://javafx.com/fxml/1">
+<fx:root type="nl.utwente.group10.ui.components.menu.FunctionMenu" xmlns:fx="http://javafx.com/fxml/1">
     <BorderPane styleClass="menu, block">
     <padding><Insets top="10" right="10" bottom="10" left="10"/></padding>
         <top>
